refactor(map): extract the first-fix decisions from the page
Importing the page from its test pulled 698 uncovered UI lines into the
coverage surface and dropped line coverage from 81.1% to 65.2%, below the
78% CI floor. The two pure decisions now live in their own file, fully
covered, and the page stays out of the report.
